Split "anylytics" part from the "generation" part.
This way we can convert commonjs modules into different kinds of target
outputs, like browser, AMD, Qt, ES6 modules etc.

Everything is a bit broken now, like watch, source-maps, compress, pipe
and so on.

I'm also using esprima / escodegen now, the AST is like a 1000 times easier
to work it, it just works, instead of other Walker.Transform stuff ujs uses.

Only thing is that their source-map support isn't that great, but that should 
be fixed soon.

I added also Qt as ouput, yesterday I did some experimenting with it, with
including prime in a QML app, and the importing is slightly different, 
because I couldn't find a global scope, so the nicest way is to have named
modules as "var name = require('...')" or something.

Basically, this commit is a start, but still needs some drastic cleaning up and
reintroducing of existing features.