Now we're able to write CommonJS modules to AMD modules with wrapup!
Needs some testing obviously, the output folder isn't nicely configurable yet
either.

What's really nice is that it puts the dependencies in the AMD dependency array
and in the function parameters, and that it replaces the require('..') calls
with the actual parameters, that are added to the function.

For example the output of:

	var a = require("./e")

Is:

	define([
		'module',
		'exports',
		'./e'
	], function (module, exports, _0) {
		var a = _0;
	});